feat: move heroBannerTags component  to folder and add test
