+++ src/components/molecules/DetailedMovieCard.tsx
@@ -4,7 +4,7 @@ import { Movie } from "../../types/Movie";
 import HeroBannerTitle from "../atoms/HeroBannerTitle/HeroBannerTitle";
 import HeroBannerDescription from "../atoms/HeroBannerDescription/HeroBannerDescription";
 import HeroBannerRating from "../atoms/HeroBannerRatings/HeroBannerRating";
-import HeroBannerTags from "../atoms/HeroBannerTags";
+import HeroBannerTags from "../atoms/HeroBannerTags/HeroBannerTags";
 
 /**
  * Variable image_url to get the url api of images
