+++ src/components/molecules/HeroBanner.tsx
@@ -3,7 +3,7 @@ import { FC } from "react";
 import { Grid, Card, Box } from "@mui/material";
 import { Link } from "react-router-dom";
 import HeroBannerDescription from "../atoms/HeroBannerDescription/HeroBannerDescription";
-import HeroBannerTags from "../atoms/HeroBannerTags";
+import HeroBannerTags from "../atoms/HeroBannerTags/HeroBannerTags";
 import HeroBannerTitle from "../atoms/HeroBannerTitle/HeroBannerTitle";
 import HeroBannerRating from "../atoms/HeroBannerRatings/HeroBannerRating";
 const image_url = import.meta.env.VITE_IMAGE_URL;
